requirements.txt
streamlit==1.28.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.0
python-dotenv==1.0.0
